INT-5752: Fixing activity menus in Folderview
For some reason, when section content had a tabindex,
then when you click on a menu item in the edit activity
menu, the browser would think that you clicked on
the section content.

Removed the tabindex and no longer focusing on the
section content once it is expanded.  Instead, now
informing the user via aria livelog that the section
is now opened/closed.